Navigation HomeScreen -> LeaderBoardScreen COMPLETE and Navigation HomeScreen -> Course_Manage COMPLETE
- duplicated the add button, 1 for language "AddLanguageButton", 1 for field "AddCourseButton". I modified it so that they receive some parameters and open new screens when clicked.

- Edited home to make navigation possible.

- Created 3 .kt files for course_manage
Add daily goal screen
Add field screen
Add language screen

- In navigation...
Screens.kt: Added new objects
NavigationWrapper.kt:
New Routes

Edited the "Home" route so that depending on the "Add" button that is clicked ("AddLanguageButton" or "AddCourseButton"), it takes you to "AddDailyGoalScreen" but with the path "/language" or "/course", so when you click to continue to the next screen it depends on the route.

I added "AddDailyScreen", which receives a "type" parameter which is the route that defines the home depending on the button that is tapped.